XP-877 Rename "SingleSelector" to "RadioButtons" input type - Renamed
