+++ apps/modules/features/src/main/resources/app/content-types/radiobuttons/radiobuttons.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<content-type>
+  <display-name>Radio buttons</display-name>
+  <super-type>base:structured</super-type>
+  <form>
+    <input name="ss_radio" type="RadioButtons">
+      <label>Radio buttons</label>
+      <indexed>true</indexed>
+      <occurrences minimum="0" maximum="0"/>
+      <config>
+        <options>
+          <option>
+            <label>myOption 1</label>
+            <value>o1</value>
+          </option>
+          <option>
+            <label>myOption 2</label>
+            <value>o2</value>
+          </option>
+          <option>
+            <label>myOption 3</label>
+            <value>o3</value>
+          </option>
+        </options>
+      </config>
+    </input>
+  </form>
+</content-type>
