--- apps/modules/features/src/main/resources/app/content-types/singleselectors/singleselectors.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<content-type>
-  <display-name>Single selectors</display-name>
-  <super-type>base:structured</super-type>
-  <form>
-    <input name="ss_radio" type="SingleSelector">
-      <label>Single Selector, radio</label>
-      <indexed>true</indexed>
-      <occurrences minimum="0" maximum="0"/>
-      <config>
-        <options>
-          <option>
-            <label>myOption 1</label>
-            <value>o1</value>
-          </option>
-          <option>
-            <label>myOption 2</label>
-            <value>o2</value>
-          </option>
-          <option>
-            <label>myOption 3</label>
-            <value>o3</value>
-          </option>
-        </options>
-        <selector-type>RADIO</selector-type>
-      </config>
-    </input>
-    <input name="ss_dropdown" type="SingleSelector">
-      <label>Single Selector, dropdown</label>
-      <indexed>true</indexed>
-      <occurrences minimum="0" maximum="0"/>
-      <config>
-        <options>
-          <option>
-            <label>myOption 1</label>
-            <value>o1</value>
-          </option>
-          <option>
-            <label>myOption 2</label>
-            <value>o2</value>
-          </option>
-          <option>
-            <label>myOption 3</label>
-            <value>o3</value>
-          </option>
-        </options>
-        <selector-type>DROPDOWN</selector-type>
-      </config>
-    </input>
-    <input name="ss_combobox" type="SingleSelector">
-      <label>Single Selector, combo-box</label>
-      <indexed>true</indexed>
-      <occurrences minimum="0" maximum="0"/>
-      <config>
-        <options>
-          <option>
-            <label>myOption 1</label>
-            <value>o1</value>
-          </option>
-          <option>
-            <label>myOption 2</label>
-            <value>o2</value>
-          </option>
-          <option>
-            <label>myOption 3</label>
-            <value>o3</value>
-          </option>
-        </options>
-        <selector-type>COMBOBOX</selector-type>
-      </config>
-    </input>
-  </form>
-</content-type>
