+++ modules/admin-ui/src/main/resources/web/admin/common/js/_module.ts
@@ -93,7 +93,7 @@
 ///<reference path='form/inputtype/_module.ts' />
 ///<reference path='form/inputtype/support/_module.ts' />
 ///<reference path='form/inputtype/combobox/_module.ts' />
-///<reference path='form/inputtype/singleselector/_module.ts' />
+///<reference path='form/inputtype/radiobuttons/_module.ts' />
 ///<reference path='form/inputtype/text/tiny/_module.ts' />
 ///<reference path='form/inputtype/text/_module.ts' />
 ///<reference path='form/inputtype/checkbox/_module.ts' />
