+++ modules/admin-ui/src/main/resources/web/admin/common/js/form/inputtype/radiobuttons/_module.ts
@@ -1,0 +1,1 @@
+///<reference path='RadioButtons.ts' />
