--- modules/admin-ui/src/main/resources/web/admin/common/js/form/inputtype/singleselector/_module.ts
@@ -1,1 +1,0 @@
-///<reference path='SingleSelector.ts' />
