--- modules/admin-ui/src/main/resources/web/admin/common/styles/api/form/inputtype/single-selector.less
@@ -1,22 +1,0 @@
-.single-selector .combobox {
-
-  display: block;
-  input {
-    height: 33px;
-    padding: 4px 37px 4px 8px;
-  }
-
-  .selected-options {
-    float: none;
-
-    .selected-option {
-      width: 100%;
-      padding-right: 0;
-
-      .remove {
-        font-size: 9px;
-      }
-    }
-  }
-
-}
