+++ modules/admin-ui/src/main/resources/web/admin/common/styles/styles.less
@@ -88,7 +88,6 @@
 @import "api/app/wizard/wizard-step-form";
 @import "api/app/wizard/wizard-step-navigator";
 @import "api/form/input/input-occurrence-view";
-@import "api/form/inputtype/single-selector";
 @import "api/form/inputtype/combo-box";
 @import "api/form/inputtype/text/tinymce";
 @import "api/form/form-item-set-occurrence-view";
