+++ modules/core-api/src/test/java/com/enonic/xp/form/inputtype/RadioButtonsTest.java
@@ -5,17 +5,17 @@ import org.junit.Test;
 import com.enonic.xp.data.PropertyTree;
 import com.enonic.xp.form.BreaksRequiredContractException;
 
-public class SingleSelectorTest
+public class RadioButtonsTest
 {
     @Test(expected = BreaksRequiredContractException.class)
     public void breaksRequiredContract_textLine_which_is_empty_throws_exception()
     {
-        new SingleSelector().checkBreaksRequiredContract( new PropertyTree().setString( "myText", "" ) );
+        new RadioButtons().checkBreaksRequiredContract( new PropertyTree().setString( "myText", "" ) );
     }
 
     @Test(expected = BreaksRequiredContractException.class)
     public void breaksRequiredContract_textLine_which_is_blank_throws_exception()
     {
-        new SingleSelector().checkBreaksRequiredContract( new PropertyTree().setString( "myText", " " ) );
+        new RadioButtons().checkBreaksRequiredContract( new PropertyTree().setString( "myText", " " ) );
     }
 }
